{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "0.1.0"
description = "{{cookiecutter.project_short_description}}"
authors = ["{{cookiecutter.full_name}} <{{cookiecutter.email}}>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.1"
uvicorn = "0.23.2"
pydantic = "^2.4.1"
pydantic-settings = "^2.0.3"
requests = "^2.22.0"
loguru = "^0.7.0"
joblib = "^1.2.0"
scikit-learn = "^1.2.0"
numpy = "^1.23.5"
opencv-python = "^4.9.0"
pillow = "^10.3.0"
ultralytics = "^8.2.0"
python-multipart = "^0.0.5"
tensorflow = "^2.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
black = {version = "^24.3", allow-prereleases = true}
autopep8 = "^2.0.0"
ipdb = "^0.13.0"
pylint = "^3.0.0"

[tool.poetry.group.aws]
optional = true

[tool.poetry.group.aws.dependencies]
magnum = "^0.17.0"
boto3 = "^1.20.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''
